backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2

pymupdf==1.24.10
python-docx==1.1.2

slowapi==0.1.9

scikit-learn==1.5.1
spacy==3.6.0
numpy==2.0.2

keybert==0.8.5
sentence-transformers==3.0.1
torch>=2.2.0
